feat / Electric Skill 1

+++ Assets/02_Scripts/Battle/Buff/Buff.cs
@@ -3,7 +3,7 @@ using UnityEngine;
 public abstract class Buff
 {
     public GameObject target;
-    public short remainTurn;
+    public int remainTurn;
 
     public abstract void TimeCheck();
 
@@ -20,7 +20,7 @@ public abstract class Buff
 
     public abstract bool StartEffect();
 
-    public abstract void SetBuff(short turn, GameObject target, short attribute = 0);
+    public abstract void SetBuff(int turn, GameObject target, int attribute = 0);
 
     public void AddBuff(GameObject target)
     {
